pyproject.toml
[project]
name = "hw2"
version = "0.1.0"
description = "Homework 2. Topic: gRPC service for ML model."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "joblib>=1.5.2",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "ruff>=0.14.7",
    "scikit-learn>=1.7.2",
]
